Added warnings for running instances

+++ launcher/ui/pages/instance/ExternalResourcesPage.cpp
@@ -250,6 +250,16 @@ void ExternalResourcesPage::removeItem()
 
 void ExternalResourcesPage::removeItems(const QItemSelection& selection)
 {
+    if (m_instance != nullptr && m_instance->isRunning()) {
+        auto response = CustomMessageBox::selectable(this, "Confirm Delete",
+                                                     "If you remove this resource while the game is running it may crash your game.\n"
+                                                     "Are you sure you want to do this?",
+                                                     QMessageBox::Warning, QMessageBox::Yes | QMessageBox::No, QMessageBox::No)
+                            ->exec();
+
+        if (response != QMessageBox::Yes)
+            return;
+    }
     m_model->deleteResources(selection.indexes());
 }
 
@@ -261,6 +271,16 @@ void ExternalResourcesPage::enableItem()
 
 void ExternalResourcesPage::disableItem()
 {
+    if (m_instance != nullptr && m_instance->isRunning()) {
+        auto response = CustomMessageBox::selectable(this, "Confirm disable",
+                                                     "If you disable this resource while the game is running it may crash your game.\n"
+                                                     "Are you sure you want to do this?",
+                                                     QMessageBox::Warning, QMessageBox::Yes | QMessageBox::No, QMessageBox::No)
+                            ->exec();
+
+        if (response != QMessageBox::Yes)
+            return;
+    }
     auto selection = m_filterModel->mapSelectionToSource(ui->treeView->selectionModel()->selection());
     m_model->setResourceEnabled(selection.indexes(), EnableAction::DISABLE);
 }
